chatbot-server/requirements.txt
fastapi
uvicorn
openai
pgvector
psycopg2-binary
python-dotenv
tiktoken
langchain-community>=0.0.21
langchain-openai>=0.1.1
